feat: divide firebase logic and rename general.css to global.css
