+++ src/firebaseActions/firestore.js
@@ -1,12 +1,10 @@
-import { FIREBASE_CONFIG } from "./constants.js";
+import { FIREBASE_CONFIG } from "../constants.js";
 import { initializeApp } from "https://www.gstatic.com/firebasejs/10.1.0/firebase-app.js";
 import { getFirestore, doc, getDoc, updateDoc, setDoc } from "https://www.gstatic.com/firebasejs/10.1.0/firebase-firestore.js";
-import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, updateProfile } from "https://www.gstatic.com/firebasejs/10.1.0/firebase-auth.js"
 
 const app = initializeApp(FIREBASE_CONFIG);
 
 const db = getFirestore(app);
-const auth = getAuth(app);
 
 class FirebaseFirestoreService {
     async getDoc(collectionName, docId) {
@@ -40,30 +38,4 @@ class FirebaseFirestoreService {
     };
 }
 
-class FirebaseAuthService {
-    async createUser(email, password, profileInfo) {
-        const cred = await createUserWithEmailAndPassword(auth, email, password);
-        const user = cred.user;
-
-        await updateProfile(user, profileInfo);
-
-        return user
-    }
-
-    getCurrentUser() {
-        return auth.currentUser
-    }
-
-    async signInUser(email, password) {
-        const cred = await signInWithEmailAndPassword(auth, email, password);
-        const user = cred.user
-        return user
-    }
-
-    async logoutUser() {
-        await signOut(auth)
-    }
-}
-
 export const firebaseFirestore = new FirebaseFirestoreService();
-export const firebaseAuth = new FirebaseAuthService();
